+++ vtm/src/org/oscim/tiling/source/mapfile/MapFileTileSource.java
@@ -1,7 +1,7 @@
 /*
  * Copyright 2013 mapsforge.org
  * Copyright 2013 Hannes Janetzek
- * Copyright 2016-2018 devemux86
+ * Copyright 2016-2020 devemux86
  *
  * This file is part of the OpenScienceMap project (http://www.opensciencemap.org).
  *
@@ -25,13 +25,13 @@ import org.oscim.tiling.OverzoomTileDataSource;
 import org.oscim.tiling.TileSource;
 import org.oscim.tiling.source.mapfile.header.MapFileHeader;
 import org.oscim.tiling.source.mapfile.header.MapFileInfo;
-import org.oscim.utils.IOUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.IOException;
-import java.io.RandomAccessFile;
+import java.nio.channels.FileChannel;
 
 public class MapFileTileSource extends TileSource implements IMapFileTileSource {
     private static final Logger log = LoggerFactory.getLogger(MapFileTileSource.class);
@@ -40,14 +40,14 @@ public class MapFileTileSource extends TileSource implements IMapFileTileSource 
      * Amount of cache blocks that the index cache should store.
      */
     private static final int INDEX_CACHE_SIZE = 64;
-    private static final String READ_ONLY_MODE = "r";
 
     MapFileHeader fileHeader;
     MapFileInfo fileInfo;
     IndexCache databaseIndexCache;
     boolean experimental;
     File mapFile;
-    private RandomAccessFile mInputFile;
+    FileInputStream mapFileInputStream;
+    private FileChannel inputChannel;
 
     /**
      * The preferred language when extracting labels from this tile source.
@@ -98,6 +98,10 @@ public class MapFileTileSource extends TileSource implements IMapFileTileSource 
         return true;
     }
 
+    public void setMapFileInputStream(FileInputStream fileInputStream) {
+        this.mapFileInputStream = fileInputStream;
+    }
+
     @Override
     public void setPreferredLanguage(String preferredLanguage) {
         this.preferredLanguage = preferredLanguage;
@@ -105,31 +109,38 @@ public class MapFileTileSource extends TileSource implements IMapFileTileSource 
 
     @Override
     public OpenResult open() {
-        if (!options.containsKey("file"))
+        if (mapFileInputStream == null && !options.containsKey("file"))
             return new OpenResult("no map file set");
 
         try {
-            // make sure to close any previously opened file first
-            //close();
+            // false positive: stream gets closed when the channel is closed
+            // see e.g. http://bugs.java.com/bugdatabase/view_bug.do?bug_id=4796385
+            File file = null;
+            if (mapFileInputStream != null)
+                inputChannel = mapFileInputStream.getChannel();
+            else {
+                // make sure to close any previously opened file first
+                //close();
 
-            File file = new File(options.get("file"));
+                file = new File(options.get("file"));
 
-            // check if the file exists and is readable
-            if (!file.exists()) {
-                return new OpenResult("file does not exist: " + file);
-            } else if (!file.isFile()) {
-                return new OpenResult("not a file: " + file);
-            } else if (!file.canRead()) {
-                return new OpenResult("cannot read file: " + file);
+                // check if the file exists and is readable
+                if (!file.exists()) {
+                    return new OpenResult("file does not exist: " + file);
+                } else if (!file.isFile()) {
+                    return new OpenResult("not a file: " + file);
+                } else if (!file.canRead()) {
+                    return new OpenResult("cannot read file: " + file);
+                }
+
+                FileInputStream fis = new FileInputStream(file);
+                inputChannel = fis.getChannel();
             }
-
-            // open the file in read only mode
-            mInputFile = new RandomAccessFile(file, READ_ONLY_MODE);
-            long mFileSize = mInputFile.length();
-            ReadBuffer mReadBuffer = new ReadBuffer(mInputFile);
+            long fileSize = inputChannel.size();
+            ReadBuffer readBuffer = new ReadBuffer(inputChannel);
 
             fileHeader = new MapFileHeader();
-            OpenResult openResult = fileHeader.readHeader(mReadBuffer, mFileSize);
+            OpenResult openResult = fileHeader.readHeader(readBuffer, fileSize);
 
             if (!openResult.isSuccess()) {
                 close();
@@ -137,10 +148,7 @@ public class MapFileTileSource extends TileSource implements IMapFileTileSource 
             }
             fileInfo = fileHeader.getMapFileInfo();
             mapFile = file;
-            databaseIndexCache = new IndexCache(mInputFile, INDEX_CACHE_SIZE);
-
-            // Experimental?
-            //experimental = fileInfo.fileVersion == 4;
+            databaseIndexCache = new IndexCache(inputChannel, INDEX_CACHE_SIZE);
 
             log.debug("File version: " + fileInfo.fileVersion);
             return OpenResult.SUCCESS;
@@ -164,8 +172,14 @@ public class MapFileTileSource extends TileSource implements IMapFileTileSource 
 
     @Override
     public void close() {
-        IOUtils.closeQuietly(mInputFile);
-        mInputFile = null;
+        if (inputChannel != null) {
+            try {
+                inputChannel.close();
+                inputChannel = null;
+            } catch (IOException e) {
+                log.error(e.getMessage());
+            }
+        }
         fileHeader = null;
         fileInfo = null;
         mapFile = null;